kernel: split Window into Widget and Window, rewrite most code
This splits the old Window class into Widget and Window, since regular
widgets do not need code for drawing a title bar, dragging and such.

Widgets still keep the concept of decoration, which are parts of the
widget that should be on top of children, and also define the relative
positioning for child widgets.